Prompt and CSV file for programming project.
Write a program that reads movie data from a CSV (comma separated values) file and output the data in a formatted table. The program first reads the name of the CSV file from the user. The program then reads the CSV file and outputs the contents according to the following requirements:

    Each row contains the title, rating, and all showtimes of a unique movie.
    A space is placed before and after each vertical separator ('|') in each row.
    Column 1 displays the movie titles and is left justified with a minimum of 44 characters.
    If the movie title has more than 44 characters, output the first 44 characters only.
    Column 2 displays the movie ratings and is right justified with a minimum of 5 characters.
    Column 3 displays all the showtimes of the same movie, separated by a space.

Each row of the CSV file contains the showtime, title, and rating of a movie. Assume data of the same movie are grouped in consecutive rows.

Ex: If the input of the program is: